Assign03
Working with methods1.Let’s make a program that uses methods to accomplisha task. Let’s take an array andreverse the contents of it. For example, if you have 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, it would
become 10, 9, 8, 7, 6, 5, 4, 3, 2, 1.To accomplish this, you’ll create three methods: one to create the array, one to reverse thearray, and one to print the array at the end.YourMainmethod will look something like this:static void Main(string[] args) {int[] numbers=GenerateNumbers();Reverse(numbers);PrintNumbers(numbers);}TheGenerateNumbersmethod should return an array of10 numbers. (For bonus points,change the method to allow the desired length to be passed in, instead of just alwaysbeing 10.)ThePrintNumbersmethodshoulduseafororforeachlooptoprintouteachiteminthearray.TheReversemethodwillbethehardest.Giveitatryandseewhatyoucanmakehappen. If you getstuck, here’s a couple of hints:Hint #1:To swap two values, you will need to placethe value of one variable in a temporarylocation to make the swap:// Swapping a and b.inta=3;intb=5;inttemp=a;a=b;b=temp;Hint #2:Getting the right indices to swap can be achallenge. Use aforloop, starting at 0and going up to the length of the array / 2. The number you use in theforloop will be theindex of the first number to swap, and the other one will be the length of the array minusthe index minus 1. This is to account for the fact that the array is 0-based. So basically,you’ll be swappingarray[index]witharray[arrayLength– index – 1].2.The Fibonacci sequence is a sequence of numberswhere the first two numbers are 1 and  1,and every other number in the sequence after it is the sum of the two numbers before it.  Sothe third number is 1 + 1, which is 2. The fourth number is the 2nd number plus the 3rd,which is 1 + 2. So the fourth number is 3. The 5th number is the 3rd number plus the 4th
number: 2 + 3 = 5. This keeps going forever.The first few numbers of the Fibonacci sequence are: 1, 1, 2, 3, 5, 8, 13, 21, 34, 55, ...Because one number is defined by the numbers before it, this sets up a perfectopportunity for using recursion.Your mission, should you choose to accept it, is to create a method calledFibonacci, whichtakes in a number and returns that number of the Fibonacci sequence. So if someone callsFibonacci(3), it would return the 3rd number in theFibonacci sequence, which is 2. Ifsomeone callsFibonacci(8), it would return 21.In yourMainmethod, write code to loop through thefirst 10 numbers of the Fibonaccisequence and print them out.Hint #1:Start with your base case. We know that ifit is the 1st or 2nd number, the value willbe 1.Hint #2:For every other item, how is it defined interms of the numbers before it? Can youcome up with an equation or formula that calls theFibonaccimethod again?Designing and Building Classes using object-oriented principles1.Write a program that that demonstrates use of fourbasic principles ofobject-oriented  programming /Abstraction/, /Encapsulation/, /Inheritance/ and/Polymorphism/.2.Use /Abstraction/ to define different classes foreach person type such as Studentand  Instructor. These classes should have behavior for that type of person.3.Use /Encapsulation/ to keep many details privatein each class.4.Use /Inheritance/ by leveraging the implementationalready created in the Personclass to  save code in Student and Instructor classes.5.Use /Polymorphism/ to create virtual methods thatderived classes could override tocreate  specific behavior such as salary calculations.6.Make sure to create appropriate /interfaces/ suchas ICourseService, IStudentService,IInstructorService, IDepartmentService, IPersonService, IPersonService (should haveperson specific methods). IStudentService, IInstructorService should inherit fromIPersonService.PersonCalculate Age of the PersonCalculate the Salary of the person, Use decimal for salarySalary cannot be negative numberCan have multiple Addresses, should have method to get addressesInstructor
Belongs to one Department and he can be Head of the DepartmentInstructor will have added bonus salary based on his experience, calculate hisyears of experience based on Join DateStudentCan take multiple coursesCalculate student GPA based on grades for coursesEach course will have grade from A to FCourseWill have list of enrolled studentsDepartmentWill have one Instructor as headWill have Budget for school year (start and end Date Time)Will offer list of courses7.Try creating the two classes below, and make asimple program to work with them, asdescribed belowCreate a Color class:On a computer, colors are typically represented with a red, green, blue, and alpha(transparency) value, usually in the range of 0 to 255. Add these as instance variables.A constructor that takes a red, green, blue, and alpha value.A constructor that takes just red, green, and blue, while alpha defaults to 255(opaque).Methods to get and set the red, green, blue, and alpha values from aColorinstance.A method to get the grayscale value for the color, which is the average of the red,green and  blue values.Create a Ball class:The Ball class should have instance variables for size and color (theColorclass you justcreated). Let’s also add an instance variable that keeps track of the number of times ithas been thrown.Create any constructors you feel would be useful.Create a Pop method, which changes the ball’s size to 0.Create a Throw method that adds 1 to the throw count, but only if the ball hasn’t beenpopped (has a size of 0).A method that returns the number of times the ball has been thrown.Write some code in your Main method to create a few balls, throw them around a fewtimes, pop a few, and try to throw them again, and print out the number of times that the
balls have been thrown. (Popped balls shouldn’t have changed.)Explore following topicsFieldsAccess modifiersEnumeration typesConstructorsMethodsPropertiesInheritanceInterfacesPolymorphism